hot fixes

+++ 0-setup_web_static.sh
@@ -1,12 +1,26 @@
 #!/usr/bin/env bash
-# sets up the web servers for the deployment of web_static
+# Bash script that sets up your web servers for the deployment of web_static
 
-sudo apt-get -y update
-sudo apt-get -y upgrade
-sudo apt-get -y install nginx
-sudo mkdir -p /data/web_static/releases/test /data/web_static/shared
-echo "This is a test" | sudo tee /data/web_static/releases/test/index.html
-sudo ln -sf /data/web_static/releases/test/ /data/web_static/current
-sudo chown -hR ubuntu:ubuntu /data/
-sudo sed -i '38i\\tlocation /hbnb_static/ {\n\t\talias /data/web_static/current/;\n\t}\n' /etc/nginx/sites-available/default
-sudo service nginx start
+# Install Nginx if it not installed
+if ! dpkg -s nginx &> /dev/null;
+then
+    sudo apt-get update;
+    sudo apt-get -y install nginx;
+    sudo ufw allow 'Nginx HTTP';
+fi;
+
+# Create the folder /data/ with all subfolders (recursively)
+sudo mkdir -p /data/web_static/releases/test/;
+sudo mkdir -p /data/web_static/shared/;
+
+# Create a fake HTML file /data/web_static/releases/test/index.html
+echo 'successful test http' | sudo tee /data/web_static/releases/test/index.html;
+
+# Create a symbolic link /data/web_static/current linked to the /data/web_static/releases/test/ folder
+sudo rm -rf /data/web_static/current;
+sudo ln -s /data/web_static/releases/test/ /data/web_static/current;
+sudo chown -R ubuntu:ubuntu /data/;
+
+# Update the Nginx configuration to serve the content of /data/web_static/current/ to hbnb_static
+sudo sed -i '/^.*server_name _;$/a \\n\tlocation /hbnb_static/ {\n\t\talias /data/web_static/current/;\n\t}\n' /etc/nginx/sites-available/default;
+sudo service nginx restart;
